fix LD Rotary buffer, now Average LD can be used with 3sec integration...
previous Avg Eff is calculated using :

AvgLD =  sum(Speed[ 0..n]) / Sum(sinkRate[0 .. n-1] )

with "n" number of second,  Speed and SinkRate in m/s
this is wrong.

now is calculated using :

AvgLD =  sum(Speed[0 .. n]) / Sum(sinkRate[0 .. n-1] )

with "n" number of second,  Speed and SinkRate in cm/s

+++ Common/Source/Calc/LD.cpp
@@ -8,7 +8,7 @@
 
 #include "externs.h"
 
-extern void InsertLDRotary(ldrotary_s *buf, int distance, NMEA_INFO *Basic, DERIVED_INFO *Calculated);
+extern void InsertLDRotary(ldrotary_s *buf, double distance, NMEA_INFO *Basic, DERIVED_INFO *Calculated);
 extern void InsertWindRotary(windrotary_s *wbuf, double speed, double track, double altitude);
 
 
@@ -79,7 +79,7 @@ void LD(NMEA_INFO *Basic, DERIVED_INFO *Calculated)
                                 DistanceFlown,
                                 LastAlt - Calculated->NavAltitude, 0.1);
 
-      if (LastLat!=0) InsertLDRotary(&rotaryLD,(int)DistanceFlown, Basic, Calculated);
+      if (LastLat!=0) InsertLDRotary(&rotaryLD, DistanceFlown, Basic, Calculated);
       InsertWindRotary(&rotaryWind, Basic->Speed, Basic->TrackBearing, Calculated->NavAltitude); // 100103
       if (ISCAR) {
          if (DistanceFlown<300) Calculated->Odometer += DistanceFlown;
